+++ include/openpose/experimental/3d/headers.hpp
@@ -1,11 +1,8 @@
 ﻿#ifndef OPENPOSE_EXPERIMENTAL_3D_HEADERS_HPP
 #define OPENPOSE_EXPERIMENTAL_3D_HEADERS_HPP
 
-// Camera intrinsic and extrinsic parameters
+// 3d module
 #include <openpose/experimental/3d/cameraParameters.hpp>
-// Flir (Point-Grey) cameras
 #include <openpose/experimental/3d/flirReader.hpp>
-// OpenGL Renderer
-#include <openpose/experimental/3d/renderer.hpp>
 
 #endif // OPENPOSE_EXPERIMENTAL_3D_HEADERS_HPP