+++ include/openpose/utilities/flagsToOpenPose.hpp
@@ -3,6 +3,7 @@
 
 #include <openpose/core/common.hpp>
 #include <openpose/core/enumClasses.hpp>
+#include <openpose/gui/enumClasses.hpp>
 #include <openpose/pose/enumClasses.hpp>
 #include <openpose/producer/producer.hpp>
 
@@ -16,18 +17,27 @@ namespace op
 
     // Determine type of frame source
     OP_API ProducerType flagsToProducerType(const std::string& imageDirectory, const std::string& videoPath,
-                                            const std::string& ipCameraPath, const int webcamIndex);
+                                            const std::string& ipCameraPath, const int webcamIndex,
+                                            const bool flirCamera);
 
     OP_API std::shared_ptr<Producer> flagsToProducer(const std::string& imageDirectory, const std::string& videoPath,
                                                      const std::string& ipCameraPath, const int webcamIndex,
                                                      const std::string& webcamResolution = "1280x720",
                                                      const double webcamFps = 30.);
 
+    OP_API std::shared_ptr<Producer> flagsToProducer(const std::string& imageDirectory, const std::string& videoPath,
+                                                     const std::string& ipCameraPath, const int webcamIndex,
+                                                     const bool flirCamera /*= false*/,
+                                                     const std::string& webcamResolution /*= "1280x720"*/,
+                                                     const double webcamFps /*= 30.*/);
+
     OP_API std::vector<HeatMapType> flagsToHeatMaps(const bool heatMapsAddParts = false,
                                                     const bool heatMapsAddBkg = false,
                                                     const bool heatMapsAddPAFs = false);
 
     OP_API RenderMode flagsToRenderMode(const int renderFlag, const int renderPoseFlag = -2);
+
+    OP_API DisplayMode flagsToDisplayMode(const int display, const bool enabled3d);
 
     OP_API Point<int> flagsToPoint(const std::string& pointString, const std::string& pointExample = "1280x720");
 }