+++ src/openpose/wrapper/wrapperStructOutput.cpp
@@ -2,13 +2,14 @@
 
 namespace op
 {
-    WrapperStructOutput::WrapperStructOutput(const bool displayGui_, const bool guiVerbose_, const bool fullScreen_,
-                                             const std::string& writeKeypoint_, const DataFormat writeKeypointFormat_,
-                                             const std::string& writeJson_, const std::string& writeCocoJson_,
-                                             const std::string& writeImages_, const std::string& writeImagesFormat_,
-                                             const std::string& writeVideo_, const std::string& writeHeatMaps_,
+    WrapperStructOutput::WrapperStructOutput(const DisplayMode displayMode_, const bool guiVerbose_,
+                                             const bool fullScreen_, const std::string& writeKeypoint_,
+                                             const DataFormat writeKeypointFormat_, const std::string& writeJson_,
+                                             const std::string& writeCocoJson_, const std::string& writeImages_,
+                                             const std::string& writeImagesFormat_, const std::string& writeVideo_,
+                                             const std::string& writeHeatMaps_,
                                              const std::string& writeHeatMapsFormat_) :
-        displayGui{displayGui_},
+        displayMode{displayMode_},
         guiVerbose{guiVerbose_},
         fullScreen{fullScreen_},
         writeKeypoint{writeKeypoint_},